3 test automation building by each minutes

+++ src/main/java/Calculator.java
@@ -16,7 +16,7 @@ public class Calculator {
 
         return a * b;
     }
-    public  double division (double a, double b) throws  Exception{
+    public  double division (int a, int b) throws  Exception{
 
         return a / b;
     }
